old-tests/submission/W3C/domhtml/HTMLTableRowElement13.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - TABLEROW</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-tr-element.html#dom-tr-deletecell" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<table align="center" summary="Table 1">
<tbody><tr>
<th>Id</th>
<th>Name</th>
<th>Position</th>
<th>Salary</th>
</tr>
</tbody></table>
<table align="center" bgcolor="#ff0000" border="4" cellpadding="2" cellspacing="2" frame="border" id="table-1" rules="all" summary="HTML Control Table" width="680">
<caption align="top">Table Caption</caption>
<thead align="center" valign="middle">
<tr align="center" bgcolor="#00FFFF" char="*" charoff="1" valign="middle">
<th id="header-1">Employee Id</th>
<th abbr="maiden" align="center" axis="center" bgcolor="#00FFFF" colspan="1" headers="header-1" height="50" id="header-2" nowrap="nowrap" rowspan="1" scope="col" valign="middle" width="100">Employee Name</th>
<th>Position</th>
<th>Salary</th>
<th>Gender</th>
<th>Address</th>
</tr>
</thead>
<tfoot align="center" valign="middle">
<tr>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
<th>next page ...</th>
</tr>
</tfoot>
<tbody align="center" valign="middle">
<tr>
<td abbr="maiden2" align="center" axis="center" bgcolor="#FF0000" colspan="1" headers="header-2" height="50" id="Table-3" nowrap="nowrap" rowspan="1" scope="col" valign="middle" width="175">EMP0001</td>
<td headers="header-2">Margaret Martin</td>
<td>Accountant</td>
<td>56,000</td>
<td>Female</td>
<td>1230 North Ave. Dallas, Texas 98551</td>
</tr>
<tr>
<td>EMP0002</td>
<td>Martha Raynolds</td>
<td>Secretary</td>
<td>35,000</td>
<td>Female</td>
<td>1900 Dallas Road Dallas, Texas 98554</td>
</tr>
</tbody>
</table>



<script>
var domtest = async_test("The deleteCell() method deletes a cell from the current row.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var cellsnodeList;
      var testNode;
      var trNode;
      var cellNode;
      var value;
      var vcells;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("tr");
      assert_equals(nodeList.length,5);
testNode = nodeList.item(3);
      cellsnodeList = testNode.cells;

      vcells = cellsnodeList.length;

      assert_equals(vcells,6);
       trNode = cellsnodeList.item(0);
      cellNode = trNode.firstChild;

      value = cellNode.nodeValue;

      assert_equals(value,"EMP0001");
       testNode.deleteCell(0);
      testNode = nodeList.item(3);
      cellsnodeList = testNode.cells;

      vcells = cellsnodeList.length;

      assert_equals(vcells,5);
       trNode = cellsnodeList.item(0);
      cellNode = trNode.firstChild;

      value = cellNode.nodeValue;

      assert_equals(value,"Margaret Martin");
       
    domtest.done();
}));

</script>
<p>The deleteCell() method deletes a cell from the current row.   

    
    Retrieve the fourth TR element and examine the value of
    the cells length attribute which should be set to six.  
    Check the value of the first TD element.  Invoke the
    deleteCell() method which will delete a cell from the current row.
    Check the value of the cell at the zero index and also check
    the number of cells which should now be five.</p>
<div id="log"></div>
</body></html>
